move ApplicationPackage to separate file
